+++ packages/backend/src/queue/types.ts
@@ -48,10 +48,13 @@ export type EndedPollNotificationJobData = {
 };
 
 export type WebhookDeliverJobData = {
+	type: string;
 	content: unknown;
 	webhookId: Webhook['id'];
 	to: string;
 	secret: string;
+	createdAt: number;
+	eventId: string;
 };
 
 export type ThinUser = {