+++ packages/backend/src/services/following/reject.ts
@@ -115,8 +115,7 @@ async function publishUnfollow(followee: Both, follower: Local) {
 
 	const webhooks = (await getActiveWebhooks()).filter(x => x.userId === follower.id && x.on.includes('unfollow'));
 	for (const webhook of webhooks) {
-		webhookDeliver(webhook, {
-			type: 'unfollow',
+		webhookDeliver(webhook, 'unfollow', {
 			user: packedFollowee,
 		});
 	}